Move AttachmentExt to extensions package
