+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/actions/AttachmentActionsBottomSheetDialog.kt
@@ -31,9 +31,9 @@ import com.infomaniak.mail.data.models.Attachment
 import com.infomaniak.mail.databinding.BottomSheetAttachmentActionsBinding
 import com.infomaniak.mail.ui.MainViewModel
 import com.infomaniak.mail.ui.main.SnackbarManager
-import com.infomaniak.mail.utils.AttachmentIntentUtils.AttachmentIntentType
-import com.infomaniak.mail.utils.AttachmentIntentUtils.executeIntent
-import com.infomaniak.mail.utils.AttachmentIntentUtils.openAttachment
+import com.infomaniak.mail.utils.extensions.AttachmentExt.AttachmentIntentType
+import com.infomaniak.mail.utils.extensions.AttachmentExt.executeIntent
+import com.infomaniak.mail.utils.extensions.AttachmentExt.openAttachment
 import com.infomaniak.mail.utils.PermissionUtils
 import com.infomaniak.mail.utils.navigateToDownloadProgressDialog
 import dagger.hilt.android.AndroidEntryPoint
